@@ -66,20 +66,20 @@
 		"@babel/core": "^7.16.0",
 		"@babel/preset-env": "^7.16.0",
 		"@bootcamp-project/bable-config": "^0.0.5",
-		"@bootcamp-project/eslint-config": "^0.0.3",
-		"@bootcamp-project/istanbul-config": "^0.0.5",
-		"@bootcamp-project/jest-config": "^0.0.5",
-		"@bootcamp-project/prettier-config": "^0.0.3",
+		"@bootcamp-project/eslint-config": "^0.1.1",
+		"@bootcamp-project/istanbul-config": "^0.1.0",
+		"@bootcamp-project/jest-config": "^0.1.1",
+		"@bootcamp-project/prettier-config": "^0.1.1",
 		"@bootcamp-project/svelte-components": "^0.0.20",
 		"@bootcamp-project/svelte-config": "^0.0.13",
-		"@bootcamp-project/typescript-config": "^0.0.24",
+		"@bootcamp-project/typescript-config": "^0.1.0",
 		"@testing-library/jest-dom": "^5.15.0",
 		"@testing-library/svelte": "^3.0.3",
 		"@types/copy-webpack-plugin": "^8.0.1",
 		"@types/csp-html-webpack-plugin": "^3.0.2",
-		"@types/jest": "^27.0.2",
+		"@types/jest": "^27.0.3",
 		"@types/mini-css-extract-plugin": "^2.3.0",
-		"@types/node": "^16.10.2",
+		"@types/node": "^16.11.9",
 		"@types/speed-measure-webpack-plugin": "^1.3.4",
 		"@types/tailwindcss": "^2.2.1",
 		"@types/webpack-dev-server": "^4.1.0",
@@ -110,7 +110,7 @@
 		"rimraf": "^3.0.2",
 		"speed-measure-webpack-plugin": "^1.5.0",
 		"style-loader": "^3.3.0",
-		"svelte": "^3.42.2",
+		"svelte": "^3.44.2",
 		"svelte-check": "^2.2.10",
 		"svelte-hot-loader": "^1.5.0",
 		"svelte-jester": "^2.1.5",
@@ -131,7 +131,7 @@
 		"workbox-window": "^6.4.1"
 	},
 	"dependencies": {
-		"ipfs-core": "^0.12.0",
+		"ipfs-core": "^0.12.1",
 		"libp2p-websockets": "^0.16.2"
 	}
 }
