@@ -1,4 +1,4 @@
 'use strict';
 
-const PrettierConfig = require('@bootcamp-project/prettier-config/prettier.config')
+const PrettierConfig = require('@bootcamp-project/prettier-config/prettier.svelte.config')
 module.exports = PrettierConfig
