@@ -1,4 +1,4 @@
 'use strict';
 
-const JestConfig = require('@bootcamp-project/jest-config/jest.config')
+const JestConfig = require('@bootcamp-project/jest-config/jest.svelte.config')
 module.exports = JestConfig
